Make ratedE and storedE base units (#35)
* Fixed reverse resolvers from DER related resolvers
* Convert ratedE and storedE to base units of Wh
* Add changeset for converting battery_unit kWh values to Wh
* Add script to generate an SQL dump of the schema based off the current version

+++ src/main/java/com/zepben/evolve/cim/iec61970/base/wires/generation/production/BatteryUnit.kt
@@ -12,12 +12,12 @@ package com.zepben.evolve.cim.iec61970.base.wires.generation.production
  * An electrochemical energy storage device.
  *
  * @property batteryState The current state of the battery (charging, full, etc.).
- * @property ratedE Full energy storage capacity of the battery. The attribute shall be a positive value.
- * @property storedE Amount of energy currently stored. The attribute shall be a positive value or zero and lower than [BatteryUnit.ratedE].
+ * @property ratedE Full energy storage capacity of the battery in watt hours (Wh). The attribute shall be a positive value.
+ * @property storedE Amount of energy currently stored in watt hours (Wh). The attribute shall be a positive value or zero and lower than [BatteryUnit.ratedE].
  */
 class BatteryUnit(mRID: String = "") : PowerElectronicsUnit(mRID) {
 
     var batteryState: BatteryStateKind = BatteryStateKind.UNKNOWN
-    var ratedE: Double = 0.0
-    var storedE: Double = 0.0
+    var ratedE: Long = 0
+    var storedE: Long = 0
 }
